move getTimeAgo to team file
