+++ components/product/Review.mkdevs.tsx
@@ -1,6 +1,27 @@
 import Icon from "$store/components/ui/Icon.tsx";
 import { StateUpdater, useMemo, useState } from "preact/hooks";
-import { getTimeAgo } from "$store/sdk/time.ts";
+
+export function getTimeAgo(date: Date): string {
+  const seconds = Math.floor((new Date().getTime() - date.getTime()) / 1000);
+  const intervals = [
+    { label: "year", seconds: 31536000 },
+    { label: "month", seconds: 2592000 },
+    { label: "week", seconds: 604800 },
+    { label: "day", seconds: 86400 },
+    { label: "hour", seconds: 3600 },
+    { label: "minute", seconds: 60 },
+  ];
+
+  for (let i = 0; i < intervals.length; i++) {
+    const interval = intervals[i];
+    const count = Math.floor(seconds / interval.seconds);
+    if (count >= 1) {
+      return `${count} ${interval.label}${count > 1 ? "s" : ""} ago`;
+    }
+  }
+
+  return "just now";
+}
 
 export interface Review {
   id: string;
